refactor(npc): split ticks.rs by tier and subsystem (TD-026)
Hub ticks.rs re-exports public fns (incl. Tier-2 fns consumed by parish-core
mint_tier2_gossip); logic moved to ticks/{tier1,tier2,tier3,gossip,prompt,truncate}.rs.
Pre-verify snapshot from interrupted run; finish pass brings to green.

Part of #1200 (TD-026).
